daydialog is separated
